+++ apps/listener/app/module/common/utils/prefixWalletCss.ts
@@ -1,9 +1,0 @@
-/**
- * Prefixes a CSS class with the wallet global CSS.
- *
- * @param name The name of the CSS class.
- * @returns The prefixed CSS class.
- */
-export function prefixWalletCss(name: string) {
-    return `frak-wallet-${name}`;
-}