+++ packages/wallet-shared/src/sharing/index.ts
@@ -27,8 +27,6 @@ export type {
 } from "./hooks/useSharingPageController";
 export { useSharingPageController } from "./hooks/useSharingPageController";
 // Icons
-export { Copy as CopyIcon } from "./icons/Copy";
-export { Share as ShareIcon } from "./icons/Share";
 // Query Keys
 export { sharingKey } from "./queryKeys";
 // Utils