+++ sdk/core/src/types/index.ts
@@ -45,14 +45,6 @@ export type {
     DisplaySharingPageResultType,
     SharingPageProduct,
 } from "./rpc/displaySharingPage";
-export type {
-    DisplayEmbeddedWalletParamsType,
-    DisplayEmbeddedWalletResultType,
-    EmbeddedViewActionReferred,
-    EmbeddedViewActionSharing,
-    LoggedInEmbeddedView,
-    LoggedOutEmbeddedView,
-} from "./rpc/embedded";
 export type { SendInteractionParamsType } from "./rpc/interaction";
 export type {
     ConditionGroup,